Refactored Code, started writing tests
