+++ src/Maestro/SubscriptionActorService/PullRequestActor.cs
@@ -199,6 +199,7 @@ namespace SubscriptionActorService
             ActionRunner = actionRunner;
             SubscriptionActorFactory = subscriptionActorFactory;
             Logger = loggerFactory.CreateLogger(GetType());
+            PullRequestDescriptionBuilder = new PullRequestDescriptionBuilder(loggerFactory);
         }
 
         public ILogger Logger { get; }
@@ -210,6 +211,7 @@ namespace SubscriptionActorService
         public IRemoteFactory DarcRemoteFactory { get; }
         public IActionRunner ActionRunner { get; }
         public IActorProxyFactory<ISubscriptionActor> SubscriptionActorFactory { get; }
+        public PullRequestDescriptionBuilder PullRequestDescriptionBuilder { get; }
 
         public async Task TrackSuccessfulAction(string action, string result)
         {
@@ -867,23 +869,25 @@ namespace SubscriptionActorService
             // non-coherency updates is 1 then combine coherency updates with those.
             // Otherwise, put all coherency updates in a separate commit.
             bool combineCoherencyWithNonCoherency = (nonCoherencyUpdates.Count == 1);
+            int startingReferenceId = 1;
+          
             foreach ((UpdateAssetsParameters update, List<DependencyUpdate> deps) in nonCoherencyUpdates)
             {
                 var message = new StringBuilder();
                 List<DependencyUpdate> dependenciesToCommit = deps;
                 await CalculateCommitMessage(update, deps, message);
-
+                Build build = await GetBuildAsync(update.BuildId);
 
                 if (combineCoherencyWithNonCoherency && coherencyUpdate.update != null)
                 {
                     await CalculateCommitMessage(coherencyUpdate.update, coherencyUpdate.deps, message);
-                    await CalculatePRDescription(coherencyUpdate.update, coherencyUpdate.deps, null, description);
+                    startingReferenceId += PullRequestDescriptionBuilder.CalculatePRDescription(coherencyUpdate.update, coherencyUpdate.deps, null, description, build, startingReferenceId);
                     dependenciesToCommit.AddRange(coherencyUpdate.deps);
                 }
 
                 List<GitFile> committedFiles = await remote.CommitUpdatesAsync(targetRepository, newBranchName, remoteFactory,
                     dependenciesToCommit.Select(du => du.To).ToList(), message.ToString());
-                await CalculatePRDescription(update, deps, committedFiles, description);
+                startingReferenceId += PullRequestDescriptionBuilder.CalculatePRDescription(coherencyUpdate.update, coherencyUpdate.deps, null, description, build, startingReferenceId);
             }
 
             // If the coherency update wasn't combined, then
@@ -891,8 +895,9 @@ namespace SubscriptionActorService
             if (!combineCoherencyWithNonCoherency && coherencyUpdate.update != null)
             {
                 var message = new StringBuilder();
+                Build build = await GetBuildAsync(coherencyUpdate.update.BuildId);
                 await CalculateCommitMessage(coherencyUpdate.update, coherencyUpdate.deps, message);
-                await CalculatePRDescription(coherencyUpdate.update, coherencyUpdate.deps, null, description);
+                startingReferenceId += PullRequestDescriptionBuilder.CalculatePRDescription(coherencyUpdate.update, coherencyUpdate.deps, null, description, build, startingReferenceId);
 
                 await remote.CommitUpdatesAsync(targetRepository, newBranchName, remoteFactory,
                     coherencyUpdate.deps.Select(du => du.To).ToList(), message.ToString());
@@ -949,164 +954,6 @@ namespace SubscriptionActorService
             }
 
             message.AppendLine();
-        }
-
-        /// <summary>
-        ///     Calculate the PR description for an update.
-        /// </summary>
-        /// <param name="update">Update</param>
-        /// <param name="deps">Dependencies updated</param>
-        /// <param name="description">PR description string builder.</param>
-        /// <returns>Task</returns>
-        /// <remarks>
-        ///     Because PRs tend to be live for short periods of time, we can put more information
-        ///     in the description than the commit message without worrying that links will go stale.
-        /// </remarks>
-        private async Task CalculatePRDescription(UpdateAssetsParameters update, List<DependencyUpdate> deps, List<GitFile> committedFiles, StringBuilder description)
-        {
-            //Find the Coherency section of the PR description
-            if (update.IsCoherencyUpdate)
-            {
-                string sectionStartMarker = $"[marker]: <> (Begin:Coherency Updates)";
-                string sectionEndMarker = $"[marker]: <> (End:Coherency Updates)";
-                int sectionStartIndex = RemovePRDescriptionSection(sectionStartMarker, sectionEndMarker, ref description);
-
-                var coherencySection = new StringBuilder();
-                coherencySection.AppendLine(sectionStartMarker);
-                coherencySection.AppendLine("## Coherency Updates");
-                coherencySection.AppendLine();
-                coherencySection.AppendLine("The following updates ensure that dependencies with a *CoherentParentDependency*");
-                coherencySection.AppendLine("attribute were produced in a build used as input to the parent dependency's build.");
-                coherencySection.AppendLine("See [Dependency Description Format](https://github.com/dotnet/arcade/blob/master/Documentation/DependencyDescriptionFormat.md#dependency-description-overview)");
-                coherencySection.AppendLine();
-                coherencySection.AppendLine(DependencyUpdateBegin);
-                coherencySection.AppendLine();
-                coherencySection.AppendLine("- **Coherency Updates**:");
-                foreach (DependencyUpdate dep in deps)
-                {
-                    coherencySection.AppendLine($"  - **{dep.To.Name}**: from {dep.From.Version} to {dep.To.Version} (parent: {dep.To.CoherentParentDependencyName})");
-                }
-                coherencySection.AppendLine();
-                coherencySection.AppendLine(DependencyUpdateEnd);
-                coherencySection.AppendLine();
-                coherencySection.AppendLine(sectionEndMarker);
-                description.Insert(sectionStartIndex, coherencySection.ToString());
-            }
-            else
-            {
-                string sourceRepository = update.SourceRepo;
-                Guid updateSubscriptionId = update.SubscriptionId;
-                Build build = await GetBuildAsync(update.BuildId);
-                string sectionStartMarker = $"[marker]: <> (Begin:{updateSubscriptionId})";
-                string sectionEndMarker = $"[marker]: <> (End:{updateSubscriptionId})";
-                int sectionStartIndex = RemovePRDescriptionSection(sectionStartMarker, sectionEndMarker, ref description);
-
-                var subscriptionSection = new StringBuilder();
-                subscriptionSection.AppendLine(sectionStartMarker);
-                subscriptionSection.AppendLine($"## From {sourceRepository}");
-                subscriptionSection.AppendLine($"- **Subscription**: {updateSubscriptionId}");
-                subscriptionSection.AppendLine($"- **Build**: {build.AzureDevOpsBuildNumber}");
-                subscriptionSection.AppendLine($"- **Date Produced**: {build.DateProduced.ToUniversalTime():MMMM d, yyyy h:mm:ss tt UTC}");
-                // This is duplicated from the files changed, but is easier to read here.
-                subscriptionSection.AppendLine($"- **Commit**: {build.Commit}");
-                string branch = build.AzureDevOpsBranch ?? build.GitHubBranch;
-                if (!string.IsNullOrEmpty(branch))
-                {
-                    subscriptionSection.AppendLine($"- **Branch**: {branch}");
-                }
-                subscriptionSection.AppendLine();
-                subscriptionSection.AppendLine(DependencyUpdateBegin);
-                subscriptionSection.AppendLine();
-                subscriptionSection.AppendLine($"- **Updates**:");
-
-                ReferenceLinksMap dependencyMapObject = new ReferenceLinksMap();
-
-                int referenceLinkId = 1;
-                foreach (DependencyUpdate dep in deps)
-                {
-                    if (!dependencyMapObject.ShaRangeToLinkId.ContainsKey((dep.From.Commit, dep.To.Commit)))
-                    {
-                        dependencyMapObject.ShaRangeToLinkId.Add((dep.From.Commit, dep.To.Commit), referenceLinkId++);
-                    }
-                }
-
-                foreach (DependencyUpdate dep in deps)
-                {
-                    subscriptionSection.AppendLine($"  - **{dep.To.Name}**: [from {dep.From.Version} to {dep.To.Version}][{dependencyMapObject.ShaRangeToLinkId[(dep.From.Commit, dep.To.Commit)]}]");
-                }
-
-                subscriptionSection.AppendLine();
-                for (int i = 1; i <= referenceLinkId; i++)
-                {
-                    foreach (KeyValuePair<(string, string), int> entry in dependencyMapObject.ShaRangeToLinkId)
-                    {
-                        if (entry.Value == i)
-                        {
-                            DependencyDetail to = deps.Find(d => d.To.Commit == entry.Key.Item2).To;
-                            try
-                            {
-                                subscriptionSection.AppendLine($"[{i}]: {GetChangesURI(to.RepoUri, entry.Key.Item1, entry.Key.Item2)}");
-                            }
-                            catch (ArgumentNullException e)
-                            {
-                                Logger.LogError(e, $"Failed to create SHA comparison link for dependency {to.Name} during asset update for subscription {update.SubscriptionId}");
-                            }
-                        }
-                    }
-                }
-
-                subscriptionSection.AppendLine();
-                subscriptionSection.AppendLine(DependencyUpdateEnd);
-                subscriptionSection.AppendLine();
-                UpdatePRDescriptionDueConfigFiles(committedFiles, subscriptionSection);
-
-                subscriptionSection.AppendLine();
-                subscriptionSection.AppendLine(sectionEndMarker);
-                description.Insert(sectionStartIndex, subscriptionSection.ToString());
-
-            }
-            description.AppendLine();
-        }
-
-        private int RemovePRDescriptionSection(string sectionStartMarker, string sectionEndMarker, ref StringBuilder description)
-        {
-            int sectionStartIndex = description.ToString().IndexOf(sectionStartMarker);
-            int sectionEndIndex = description.ToString().IndexOf(sectionEndMarker);
-
-            if (sectionStartIndex != -1 && sectionEndIndex != -1)
-            {
-                sectionEndIndex += sectionEndMarker.Length;
-                description.Remove(sectionStartIndex, sectionEndIndex - sectionStartIndex);
-                return sectionStartIndex;
-            }
-            // if either marker is missing, just append at end and don't remove anything
-            // from the description
-            return description.Length;
-        }
-
-        private string GetChangesURI(string repoURI, string from, string to)
-        {
-            if (repoURI == null)
-            {
-                throw new ArgumentNullException(nameof(repoURI));
-            }
-            if (from == null)
-            {
-                throw new ArgumentNullException(nameof(from));
-            }
-            if (to == null)
-            {
-                throw new ArgumentNullException(nameof(to));
-            }
-
-            string fromSha = from.Length > 7 ? from.Substring(0, 7) : from;
-            string toSha = to.Length > 7 ? to.Substring(0, 7) : to;
-
-            if (repoURI.Contains("github.com"))
-            {
-                return $"{repoURI}/compare/{fromSha}...{toSha}";
-            }
-            return $"{repoURI}/branches?baseVersion=GC{fromSha}&targetVersion=GC{toSha}&_a=files";
         }
 
         private async Task UpdatePullRequestAsync(InProgressPullRequest pr, List<UpdateAssetsParameters> updates)
